数据类型

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesEquipmentAtomOperation.java
@@ -58,7 +58,7 @@ public class IcesEquipmentAtomOperation extends BaseEntity {
     /**
      * 执行时长
      */
-    private Long eaoExecTime;
+    private Float eaoExecTime;
     /**
      * 已删除
      */
