+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesManufactureTask.java
@@ -72,11 +72,11 @@ public class IcesManufactureTask extends BaseEntity {
     /**
      * 计划产品数量
      */
-    private Long mtQtyPlan;
+    private Float mtQtyPlan;
     /**
      * 已完成产品数量
      */
-    private Long mtQtyReal;
+    private Float mtQtyReal;
     /**
      * 已删除
      */
