+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/bo/IcesEquipmentOperationBo.java
@@ -47,7 +47,7 @@ public class IcesEquipmentOperationBo extends BaseEntity {
     /**
      * 执行时长
      */
-    private Long eoExecTime;
+    private Float eoExecTime;
 
     /**
      * 已删除
