+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/vo/IcesEquipmentOperationVo.java
@@ -56,7 +56,7 @@ public class IcesEquipmentOperationVo {
      * 执行时长
      */
     @ExcelProperty(value = "执行时长")
-    private Long eoExecTime;
+    private Float eoExecTime;
 
     /**
      * 已删除
